decrease src file && bsen programming start
