+++ tms_ss/tms_ss_pot/config.yaml
@@ -1,0 +1,7 @@
+max_lrf_range: 4.0
+min_lrf_range: 0.2
+target_area: [ 0.0, 0.0, 15.0, 7.0 ]
+m_min_distance: 1000
+max_vel: 100.0
+m_sigma: 200.0
+n_of_particles: 100
