+++ tms_ss/tms_ss_pot/include/multiple_particle_filter.h
@@ -10,10 +10,9 @@ public:
 	std::vector<CPF> m_ParticleFilter;
 	CLaser* m_pLaser;
 
-
 protected:
 	int    m_max_ID;
 	double m_min_distance;
+	double m_initial_dist;
 	int    m_ID;
-
 };
