+++ tms_ss/tms_ss_pot/include/target.h
@@ -1,0 +1,21 @@
+#pragma once
+
+// CTarget
+
+class CTarget
+{
+public:
+	CTarget();
+	virtual ~CTarget();
+
+public:
+	int id;
+	double px;
+	double py;
+	int cnt;
+
+public:
+	void SetPosi(double x, double y);
+};
+
+
