--- tms_ss/tms_ss_pot/include/tms_ss_pot/Target.h
@@ -1,20 +1,0 @@
-#pragma once
-
-// CTarget 
-
-class CTarget
-{
-public:
-	CTarget();
-	virtual ~CTarget();
-
-public:
-	int id;
-	double px;
-	double py;
-	int cnt;
-
-public:
-	void SetPosi(double x, double y);
-};
-
