--- tms_ss/tms_ss_pot/launch/urg.launch
@@ -1,12 +1,0 @@
-<launch>
-  <group ns="urg0">
-    <node pkg="urg_node" name="urg_node" type="urg_node">
-    <param name="serial_port" value="/dev/ttyACM0" />
-    </node>
-  </group>
-   <group ns="urg1">
-    <node pkg="urg_node" name="urg_node" type="urg_node">
-    <param name="serial_port" value="/dev/ttyACM1" />
-    </node>
-  </group>
-</launch>
